associate vscode pages with snippets (#51)
allow "starring" of snippets to associate them with pages in a
repository.

this "simple" feature is turning into a rewrite of a lot of the code to
make happen.

also turning off most of the bot functionality to focus on vs code
extension. will be turned back on in next pull request to make work with
the new db structure.